# docs/30-tech_specs/backend/openapi.v1.yaml
...
openapi: 3.0.3
info:
  title: PixelProwlers Backend API
  version: "1.3.0"
  description: |
    API REST du backend Django PixelProwlers.
    
    ## Principes clés
    - **Privacy-first** : Aucun tracking, minimisation des données
    - **Erreurs neutres** : Pas d'info technique exposée (anti-probing)
    - **Rate limiting** : Protection anti-abus sur tous les endpoints publics
    - **No PII in logs** : Jamais d'email/message en clair dans les logs
    
    ## Authentification
    - Endpoints publics (`/health/`, `/contact/`) : Aucune auth requise
    - Endpoints Gate125 : API Key requise (future)
    - Endpoints admin : Session Django
  contact:
    name: PixelProwlers Support
    email: contact@pixelprowlers.io
    url: https://pixelprowlers.io
  license:
    name: Proprietary
    url: https://pixelprowlers.io/legal
servers:
  - url: "{protocol}://{host}"
    description: Serveur configurable par environnement
    variables:
      protocol:
        default: https
        enum:
          - http
          - https
        description: Protocole (http pour dev, https pour prod)
      host:
        default: api.pixelprowlers.com
        description: Hostname de l'API (placeholder, remplacer selon env)
...
paths:
...
  /api/v1/gate125/register/:
    post:
      operationId: registerGate125
      summary: Inscription Operation 125
      description: |
        **STATUS: Future (non implémenté)**
        
        Endpoint pour l'inscription à l'Operation 125.
        Accès restreint via API Key.
        
        **Rate limit**: 10 requêtes/heure/IP (cooldown 1h après dépassement)
        
        TODO: Définir le flow d'obtention de l'API Key
      tags:
        - Gate125
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Gate125RegisterRequest"
      responses:
        "201":
          description: Inscription enregistrée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gate125RegisterSuccess"
        "400":
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "401":
          description: API Key manquante ou invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "429":
          description: Rate limit dépassé
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
            Retry-After:
              $ref: "#/components/headers/Retry-After"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitError"

  # --------------------------------------------------------------------------
  # RESOURCES ENDPOINTS (FUTURE)
  # --------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        API Key pour endpoints protégés Gate125.
        Format: `Api-Key <KEY>`
        Header alternatif accepté: `X-API-Key: <KEY>`
        
        TODO: Définir process d'obtention de la clé

  # --------------------------------------------------------------------------
  # HEADERS
  # --------------------------------------------------------------------------
  headers:
    X-RateLimit-Limit:
      description: Nombre max de requêtes autorisées dans la fenêtre
      schema:
        type: integer
        example: 3
    X-RateLimit-Remaining:
      description: Nombre de requêtes restantes dans la fenêtre
      schema:
        type: integer
        example: 2
    X-RateLimit-Reset:
      description: Timestamp Unix de reset du rate limit
      schema:
        type: integer
        example: 1703419200
    Retry-After:
      description: Secondes avant de pouvoir réessayer
      schema:
        type: integer
        example: 300

  # --------------------------------------------------------------------------
  # SCHEMAS
  # --------------------------------------------------------------------------
  schemas:
...
    Gate125RegisterRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: Email d'inscription
        organization:
          type: string
          maxLength: 200
          description: Organisation (optionnel)
        role:
          type: string
          enum:
            - porteur_projet
            - formateur
            - accompagnant
            - autre
          description: Rôle dans le projet
    Gate125RegisterSuccess:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Inscription enregistrée."

    # === RESOURCES ===
...
    ValidationError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Données invalides"
          description: |
            Message générique (anti-probing).
            Ne révèle PAS quel champ est invalide.
    RateLimitError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Trop de requêtes"
...
    UnauthorizedError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Accès non autorisé"
          description: API Key manquante ou invalide

# ============================================================================
# NOTES
# ============================================================================
# 
# RATE LIMITS PAR ENDPOINT:
# - /api/v1/health/*      : Illimité
# - /api/v1/contact/      : 3/min/IP (cooldown 5 min)
# - /api/v1/gate125/*     : 10/h/IP (cooldown 1h)
# - /api/v1/resources/    : 60/min/IP
#
# SÉCURITÉ:
# - CORS: pixelprowlers.com uniquement (+ localhost en dev)
# - CSRF: Token cookie + header pour POST
# - HSTS: Activé en production (max-age=31536000)
# - Headers: X-Content-Type-Options, X-Frame-Options, etc.
#
# VALIDATION (appliquée côté serveur):
# - NFKC normalization sur tous les strings
# - Suppression caractères zero-width
# - Trim whitespace
# - Allowlist enum stricte pour subject/role
#
# LOGS:
# - NO PII: jamais d'email/message en clair
# - IP hashée avec salt rotatif journalier
# - Correlation ID éphémère par requête
#
# ============================================================================